feat(schema): unique tables for media and schema for deduplication
it's common that providers share a show (or in Hulu's case, have
separate entries for the same shows), so it's better to store individual
media entries per provider, and only store one rating. this somewhat
simplifies the null handling of the database, although ranks are a bit
annoying to manage.